requirements.txt
Django==2.0.5
psycopg2-binary==2.7.4
django-compressor==2.2
djangorestframework==3.7.7
django-sharding=3.0.0
